refactor: 산책일지 생성시 excrements_count 컬럼에 값 추가
- journals 타입 한 파일로 합침
