+++ backend/src/journals/journals.controller.ts
@@ -18,7 +18,7 @@ import { CreateJournalDto } from './dtos/create-journal.dto';
 import { UpdateJournalDto } from './dtos/update-journal.dto';
 import { AuthJournalGuard } from './guards/auth-journal.guard';
 import { JournalsService } from './journals.service';
-import { JournalInfoForList } from './types/journal-info.type';
+import { JournalInfoForList } from './types/journal.types';
 
 import { AccessTokenPayload } from '../auth/token/token.service';
 import { AuthDogGuard } from '../dogs/guards/auth-dog.guard';
