--- backend/src/journals/types/update-journal-data.type.ts
@@ -1,7 +1,0 @@
-export type UpdateTodayWalkTimeOperation = (current: number, operand: number) => number;
-export type UpdateDogWalkDayOperation = (current: number) => number;
-
-export interface UpdateJournalData {
-    memo: string;
-    journalPhotos: string[];
-}
